answer-generator/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
requests==2.31.0

# LLM and AI
google-generativeai==0.8.3
tenacity==9.0.0

# AWS SDK
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
